picasso-bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
Mastodon.py==1.8.1
lxml==5.2.1
